User input for recursion.
This makes it possible for the user to not have to specify recursion y/n
from start. Once a dir is encountered, if recursion y/n is not defined
then it will prompt the user.

Also removed the Caleb_Smith_Test.txt since it was only for testing
pusing to the repo.