DNA_Reader uses new ABIF_Reader (formerly Randazzo_Testing)
This commit adds the complete use of the DNA reader.
The commandline input are as follows:
INPUT OUTPUT RECURSE VERBOSE

If input is a directory, then output will be the name of the new dir
created, each file in the output directory has the name of the original
file with .xml as the extension.
if verbose is "y" then extra print statements will appear while using
the reader, this is for debugging purposes.